==> notebooks/1-pca.ipynb <==
import pandas as pd
import os
import matplotlib.pyplot as plt
import seaborn as sns

from jacksonii_analyses import vcf_parser, clustering
from jacksonii_analyses.plotting import *

from importlib import reload
reload(vcf_parser)
reload(clustering)

os.makedirs("../data/figs", exist_ok=True)

variants = vcf_parser.vcf_to_geno_df(
    "../data/var/filtered_variants.vcf.gz",
)
variants.iloc[:5, :5]  # Display the first 5 rows and columns

pops = clustering.read_populations(
    "../data/samples/populations.txt",
)

pops.loc[:, "colormap"] = pops["populations_clean"].map(map_colors)
pops.head()

pc_df = clustering.calculate_pca(
    geno_df=variants,
    n_components=10,
    pops=pops,
)
pc_df.head()

pc_df.attrs["variance_explained"]

palette = dict(zip(pops["populations_clean"], pops["colormap"]))
plt.figure(figsize=(5.5, 3.7))  # Smaller plot
sns.scatterplot(
    data=pc_df,
    x='PC1',
    y='PC2',
    hue='populations',
    style='populations',
    palette=palette,
    s=60,
    edgecolor='black',
    linewidth=0.8   
)
plt.axhline(0, color='black', linestyle='dotted', linewidth=1)
plt.axvline(0, color='black', linestyle='dotted', linewidth=1)
plt.legend(title='Population', bbox_to_anchor=(1.05, 1), loc='upper left')
plt.tight_layout()
plt.savefig("../data/figs/pca_plot_color.svg")  # Save as SVG
plt.show()

plt.figure(figsize=(5.5, 3.7))  # Smaller plot
sns.scatterplot(
    data=pc_df,
    x='PC1',
    y='PC2',
    hue='populations',
    style='populations',
    palette='grey',
    s=60,
    edgecolor='black',
    linewidth=0.8   
)
plt.axhline(0, color='black', linestyle='dotted', linewidth=1)
plt.axvline(0, color='black', linestyle='dotted', linewidth=1)
plt.legend(title='Population', bbox_to_anchor=(1.05, 1), loc='upper left')
plt.tight_layout()
plt.savefig("../data/figs/pca_plot_grey.svg")  # Save as SVG
plt.show()
